% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_d_wad.R
\name{calc_d_wad}
\alias{calc_d_wad}
\title{Calculation of weighted average density differences}
\usage{
calc_d_wad(data)
}
\arguments{
\item{data}{Data as a \code{phyloseq} object}
}
\value{
\code{calc_d_wad} adds two S4 Matrix objects to the \code{data@qsip@.Data} slot, one for differences
  in weighted average density, and the other for weighted average density values of light treatments only (to be used in
  future calculations). The row and column specifications will mirror those of the \code{phylosip}'s \code{\link{otu_table}},
  meaning if taxa are listed on the table rows, they will in the resulting S4 Matrix objects
}
\description{
Calculates differences in weighted average densities across replicates due to isotope incorporation
}
\details{
Some details about proper isotope control-treatment factoring. If weighted average densities have not been calculated
  beforehand, \code{calc_d_wad} will compute those first.
}
\examples{
 # Load in example data

 # Calculate weighted average density differences

}
\seealso{
\code{\link{calc_wad}}
}
