% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_d_waf.R
\name{calc_d_waf}
\alias{calc_d_waf}
\title{Grouping of weighted average fraction}
\usage{
calc_d_waf(data, filter = FALSE, correction = FALSE,
  offset_taxa = 0.1, separate_light = FALSE, separate_label = TRUE,
  global_light = FALSE, recalc = TRUE, rel_abund = TRUE)
}
\arguments{
\item{data}{Data as a \code{phyloseq} object}

\item{filter}{Logical vector specifying whether or not to filter taxa from the weighted average density calculation.
This will require \code{data} to have a filter applied with \code{\link{filter_qsip}}.}

\item{correction}{Logical value indicating whether or not to apply tube-level correction to labeled WAF values.}

\item{offset_taxa}{Value from 0 to 1 indicating the percentage of the taxa to utilize for calculating offset correction values.
Taxa are ordered by lowest difference in WAD values.
Default is \code{0.1} indicating 10 percent of taxa with the lowest difference in WAD values.}

\item{separate_light}{Logical value indicating whether or not WAD-light scores should be averaged across all replicate groups or not.
If \code{FALSE}, unlabeled WAD scores across all replicate groups will be averaged, creating a single molecular weight score per taxon
representing it's genetic molecular weight in the absence of isotope addition.}

\item{separate_label}{Logical value indicating whether or not WAD-label scores should be averaged across all replicate groups or not.
If \code{FALSE}, labeled WAD scores across all replicate groups will be averaged, creating a single molecular weight score per taxon
representing it's genetic molecular weight as a result of isotope addition. The default is \code{TRUE}.}

\item{global_light}{Logical value indicating whether or not to use WAD-light scores that are global averages (\emph{i.e.,} averaged across
all samples rather than averaged across any specified replicate groups). The default is \code{FALSE}.}

\item{recalc}{Logical value indicating whether or not to recalculate WAD, WAD heavy, and WAD light values or use existing values. Default is \code{TRUE}.}

\item{rel_abund}{Logical value specifying if relative abundances of taxa are to be calculated prior to calculations. The default is \code{TRUE}.
This parameter is passed to \code{calc_wad}.}
}
\value{
\code{calc_d_waf} adds three S4 Matrix class objects (which more efficiently stores sparse matrix data) to the \code{data@qsip@.Data} slot
  of weighted average fraction for each taxon at each group of replicates in the labeled and unlabeled groups. The row and column
  specifications will mirror those of the \code{phylosip}'s \code{\link{otu_table}}, meaning if taxa are listed on the table rows,
  they will in the resulting S4 Matrix class.

  If timepoint is specified through \code{data@qsip@timepoint}, then \code{calc_d_waf} will group molecular weights by timepoint as a grouping factor.
  If both \code{data@qsip@timepoint} and \code{data@qsip@rep_group} are specified, then \code{calc_d_waf} will group weighted average fractions by the
  interaction of these terms. There is no option to change this behavior except by removing the specification for \code{data@qsip@timepoint}.
  This is because calculating weighted average fraction for separate timepoints is a necessity in \code{\link{calc_pop}}
}
\description{
Groups weighted average fraction values of microbial taxa by isotope treatment
}
\details{
Some details about proper isotope control-treatment factoring. If weighted average fractions
  have not been calculated beforehand, \code{calc_d_waf} will compute those first.

  Explain consequences of different grouping actions on results.
}
\examples{
 # Load in example data


}
\seealso{
\code{\link{calc_waf}}
}
