% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mw.R
\name{calc_mw}
\alias{calc_mw}
\title{Calculation of molecular weights}
\usage{
calc_mw(data, filter = FALSE, correction = FALSE, offset_taxa = 0.1,
  separate_light = FALSE, separate_label = TRUE, recalc = TRUE)
}
\arguments{
\item{data}{Data as a \code{phyloseq} object}

\item{filter}{Logical vector specifying whether or not to filter taxa from the weighted average density calculation.
This will require \code{data} to have a filter applied with \code{\link{filter_qsip}}.}

\item{correction}{Logical value indicating whether or not to apply tube-level correction to labeled WAD values.}

\item{offset_taxa}{Value from 0 to 1 indicating the percentage of the taxa to utilize for calculating offset correction values.
Taxa are ordered by lowest difference in WAD values.
Default is \code{0.1} indicating 10 percent of taxa with the lowest difference in WAD values.}

\item{separate_light}{Logical value indicating whether or not WAD-light scores should be averaged across all replicate groups or not.
If \code{FALSE}, unlabeled WAD scores across all replicate groups will be averaged, creating a single molecular weight score per taxon
representing it's genetic molecular weight in the absence of isotope addition.}

\item{separate_label}{Logical value indicating whether or not WAD-label scores should be averaged across all replicate groups or not.
If \code{FALSE}, labeled WAD scores across all replicate groups will be averaged, creating a single molecular weight score per taxon
representing it's genetic molecular weight as a result of isotope addition. The default is \code{TRUE}.}

\item{recalc}{Logical value indicating whether or not to recalculate WAD, WAD heavy, and WAD light values or use existing values. Default is \code{TRUE}.}
}
\value{
\code{calc_mw} adds three S4 Matrix class objects (which more efficiently stores sparse matrix data) to the \code{data@qsip@.Data} slot
  of molecular weights for each taxon at each group of replicates in the labeled and unlabeled groups. The row and column
  specifications will mirror those of the \code{phylosip}'s \code{\link{otu_table}}, meaning if taxa are listed on the table rows,
  they will in the resulting S4 Matrix class.
}
\description{
Calculates molecular weights of microbial taxa due to isotope incorporation
}
\details{
Some details about proper isotope control-treatment factoring. If weighted average densities or the change in weighted average densities
  have not been calculated beforehand, \code{calc_mw} will compute those first.

  The equations for calculating the molecular weights of taxon \emph{i}, designated \eqn{M_{Lab,i}} for labeled and \eqn{M_{Light,i}}, are

  \deqn{M_{Light,i} = 0.496 \cdot G_{i} + 307.691}
  \deqn{M_{Lab,i} = \left( \frac{\Delta W}{W_{Light,i}} + 1 \right) \cdot M_{Light,i}}
}
\examples{
 # Load in example data

 # Calculate molecular weights

}
\seealso{
\code{\link{calc_wad}}, \code{\link{calc_d_wad}}
}
