% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pop.R
\name{calc_pop}
\alias{calc_pop}
\title{Calculation of population dynamics}
\usage{
calc_pop(
  data,
  tax_id = c(),
  sample_id = c(),
  wads = "wad",
  iso_trt = c(),
  isotope = c(),
  timepoint = c(),
  abund = c(),
  correction = TRUE,
  rm_outliers = TRUE,
  non_grower_prop = 0.1,
  mu = 0.6,
  nat_abund_13C = 0.01111233,
  nat_abund_15N = 0.003663004,
  nat_abund_18O = 0.002011429
)
}
\arguments{
\item{data}{Data as a long-format data.table where each row represents a taxonomic feature within a single fraction.
Typically, this is the output from the \code{calc_wad} function.}

\item{tax_id}{Column name specifying unique identifier for each taxonomic feature.}

\item{sample_id}{Column name specifying unique identifier for each replicate.}

\item{wads}{Column name specifying weighted average density values.}

\item{iso_trt}{Column name specifying a two-level categorical column indicating whether a sample has been amended with a stable isotope (i.e., is "heavy") or if
isotopic composition is at natural abundance (i.e., "light").
Any terms may be applied but care should be taken for these values.
If supplied as a factor, \code{calc_excess} will take the lowest level as the "light" treatment and the higher
level as the "heavy" treatment.
Alternatively, if supplied as a character, \code{calc_excess} will coerce the column to a factor and with the default behavior wherein the
first value in alphabetical order will be assumed to be the lowest factor level (i.e. the "light" treatment).}

\item{isotope}{Column name specifying the isotope applied to each replicate. For "heavy" samples, these values should be one of "13C", "15N", or "18O".}

\item{correction}{Whether to apply a correction to fractional enrichment values to ensure a certain proportion are positive.}

\item{rm_outlers}{Whether or not to remove fractional enrichment values that are 1.5X greater or lesser than the distance between the median
and interquartile ranges.}

\item{non_grower_group}{Fractional value applied if \code{correction == TRUE} specifying the proportion of the community in each samples assumed to be
non-growers and whose median enrichment values will be assumed to be zero. The adjustment necessary to place this median value at zero will be applied
as a correction to all enrichment values in the sample.}
}
\value{
\code{calc_pop} returns a data.table where each row represents a taxonomic feature within a single replicate.
 The following columns are produced: growth rate (\code{growth}), mortality or turnover rate (\code{mortality}).
}
\description{
Calculates population growth and death rates
}
\details{
Details to go here.
}
\examples{
 # Load in example data

 # Calculate population fluxes

}
\references{
Koch, Benjamin, \emph{et al.} 2018. Estimating taxon-specific population dynamics in diverse microbial communities.
 \emph{Ecosphere} \strong{9}.
}
\seealso{
\code{\link{calc_wad}}
}
