% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{calc_tiny_abundance}
\alias{calc_tiny_abundance}
\title{Calculate Tiny Abundance}
\usage{
calc_tiny_abundance(df, column = "t.copies.ul", multiplier = 0.5)
}
\arguments{
\item{df}{The starting qSIP dataframe}

\item{column}{The column header to adjust}

\item{multiplier}{The adjustment multiplier}
}
\description{
A function to return the adjustments to zero-value data in the specified
column.
}
