% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_waf.R
\name{calc_waf}
\alias{calc_waf}
\title{Calculation of weighted average fraction}
\usage{
calc_waf(data, filter = FALSE, pool_unlabeled = TRUE,
  calc_wvf = FALSE, rel_abund = TRUE)
}
\arguments{
\item{data}{Data as a \code{phylosip} object}

\item{filter}{Logical value specifying whether or not to filter taxa from the weighted average density calculation.
A hard filter is applied for \code{calc_wad} (meaning taxa are removed from all replicates/groups if they don't meet threshold).
Note, however, that this will have different outcomes depending on whether or not \code{filter_qsip} has been called
before calculations or not (see \code{details})}

\item{pool_unlabeled}{Logical value specifying if unlabeled replicates should be pooled together across any grouping factor prior
to filtering. If \code{TRUE} (the default), unlabeled replicates will be pooled, and any soft filtering threshold will be applied to
\emph{all} unlabeled replicates together.}

\item{calc_wvf}{Logical value specifying whether or not to calculate weighted variance of fraction (WVF).
The weighted variance of fraction of a taxon within a replicate is the complement to its weighted average.
It describes how much variance exists around every given weighted average fraction.
If \code{TRUE}, another output matrix (\code{wvd}) will be returned with identical dimensions to the weighted average
fraction output.}

\item{rel_abund}{Logical value specifying whether or not to relativize the abundances in the \code{otu_table}. Relative abundances
are calculated \emph{before} filtering takes place. If relative abundances are already calculated, or if extensive filtering of taxa has
already taken place, then this may be set to \code{FALSE}.}
}
\value{
\code{calc_waf} adds an S4 Matrix class (which more efficiently stores sparse matrix data) to the \code{.Data} slot within
  the \code{qsip} slot of the object. of weighted average fraction values for each taxon at each sample. The row and column
  specifications will mirror those of the \code{phylosip}'s \code{\link{otu_table}}, meaning if taxa are listed on the table rows,
  they will in the resulting S4 Matrix class.
}
\description{
Calculates weighted average fraction for each microbial taxa in each sample replicate
}
\details{
Filtering for \code{calc_waf} utilized \code{filter_qsip} and employs a hard filter. If filtering is specified using
  \code{filter=TRUE}, then filtering is employed on fractions \emph{only} (\emph{i.e.}, \code{filter_qsip} is implemented with
  \code{replicate=1}). In order to set more stringent hard filters, \code{filter_qsip} must be employed before calculating WAD values.
  We \emph{strongly} recommend that \code{rm_low_freq} be set as \code{TRUE}, as this should reduce the occurrence of
  negative shifts in fraction

  The weighted average fraction for taxon \emph{i} in replicate \emph{j}, designated as \eqn{W_{ij}}, is:

  \deqn{W_{ij} = \sum_{k=1}^{K} x_{jk} \cdot \left( \frac{y_{ijk}}{y_{ij}} \right)}

  Where
  \eqn{y_{ij} = \sum_{k=1}^{K} y_{ijk}}: Total abundance of taxon \emph{i} in replicate \emph{j}, by summing across all fractions (\emph{K})

  \eqn{y_{ijk} = P_{ijk} \cdot f_{jk}}: The total abundance per \eqn{\mu}L for taxon \emph{i} in density fraction
  \emph{k} of replicate \emph{i} as calculated by it's relative abundance, \eqn{P_{ijk}} in that fraction multiplied by the
  total abundance of DNA or specific amplicons in that fraction, \eqn{f_{jk}}

  \eqn{x_{jk}}: Density of fraction \emph{k} of replicate \emph{j} in (g cm\eqn{^-3})
}
\examples{
 # Load in example data

 # Calculate weighted average densities


}
\seealso{
\code{\link{filter_qsip}}
}
