% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_message.R
\name{comparison_message}
\alias{comparison_message}
\title{Comparison Message}
\usage{
comparison_message(
  X.light,
  X.heavy,
  boot.out.light,
  boot.out.heavy,
  var = "trt.code"
)
}
\arguments{
\item{X.light}{data frame with data for treatment 1, the unlabeled, or 'light', treatment}

\item{X.heavy}{data frame with data for treatment 2, the labeled, or 'heavy', treatment}

\item{boot.out.light}{list containing output from boot.WAD.func or boot.TUBE.func for the 'light' treatment (corresponding to X.light)}

\item{boot.out.heavy}{list containing output from boot.WAD.func or boot.TUBE.func for the 'heavy' treatment (corresponding to X.heavy)}

\item{var}{variable name for the treatment ID columns in the data frames; default is "trt.code"}
}
\value{
warning message listing replicates in either of the 2 treatments (does not include the reference group) that had no occurrences (i.e., all y values for a rep are 0); value for message is "none" when all replicates are present (single character value)
}
\description{
Returns a warning message about which replicate(s) are missing data when comparing two treatments
}
\details{
Written by Ben Koch & Natasja van Gestel
}
\note{
number of reps need not be equal between the two groups
}
