% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_wad_calc.R
\name{diff_wad_calc}
\alias{diff_wad_calc}
\title{Calculate WAD Differences}
\usage{
diff_wad_calc(
  X1,
  X2,
  boot.out1,
  boot.out2,
  var = "trt.code",
  CI = 0.9,
  draws = 1000,
  tailed.test = 2
)
}
\arguments{
\item{X1}{data frame with data for treatment 1}

\item{X2}{data frame with data for treatment 2}

\item{boot.out1}{list containing output from boot.WAD.func or boot.TUBE.func for treatment 1 (corresponding to X1; the 'light' treatment if a 1-tailed permutation test is specified)}

\item{boot.out2}{list containing output from boot.WAD.func or boot.TUBE.func for treatment 2 (corresponding to X2; the 'heavy' treatment if a 1-tailed permutation test is specified)}

\item{var}{variable name for the treatment ID columns in the data frames}

\item{CI}{confidence interval (0-1); default is 0.90}

\item{draws}{number of iterations for the permutation test}

\item{tailed.test}{specifies whether a 2-tailed or 1-tailed permutation test should be performed in calculating the p-value; default is 2; note: if 1-tailed is specified, then the function assumes group 2 (X2) is the 'heavy' treatment and group 1 (X1) is the light treatment}
}
\value{
A list with
\itemize{
  \item boot.wads1: vector of bootstrapped WADs for treatment group 1 (X1)
  \item boot.wads2: vector of bootstrapped WADs for treatment group 2 (X2)
  \item boot.diffs: vector of bootstrapped difference in WAD between group 2 and group 1 (calculated as group2 - group1)
  \item obs.wads.by.group: data frame of mean (across reps) of the observed WAD for each group
  \item boot.wads.by.group: data frame of mean, median, CI-lower, and CI-upper of the bootstrapped WADs for each group
  \item obs.diff: observed difference in WADs between treatments (i.e. subtracting mean WAD -- across reps -- for group 1 from that for group 2)
  \item boot.diffs.mean: mean of the bootstrapped differences in WAD between group 2 and group 1
  \item boot.diffs.median: median of the bootstrapped differences in WAD between group 2 and group 1
  \item boot.diffs.CI: upper and lower confidence intervals (defined by argument 'CI') of the bootstrapped differences in WAD between group 2 and group 1
  \item p.value: P-value (obtained by permutation test); the probability of observing a difference between treatments that is as large or larger than the observed difference, if the null hypothesis (that the treatments have equal WADs) is true
  \item message: warning message listing replicates in either group that had no occurrences (i.e., all y values for a rep are 0); value for message is "none" when all replicates are present (single character value)
}
}
\description{
Given two treatments, returns statistics to assess whether the means of
weighted average density (WAD) differ significantly between treatments.
}
\details{
Written by Ben Koch & Natasja van Gestel
}
\note{
requires the functions 'WAD.func', 'boot.WAD.func', & 'comparison.message'

names of the x, y, replicate ID, and treatment ID columns in the two data frames must match exactly

number of reps need not be equal between the two groups

the test statistic for the P-value calculation is the difference in means of the 2 groups (calculated as group2 - group1)

the P-value estimates the probability that the observed difference (group2-group1) is as extreme or more extreme than would be expected if group identities were randomized
}
