% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_qsip}
\alias{example_qsip}
\title{Example qSIP dataset}
\format{
A data.table with 14387 rows and 18 variables
}
\source{
<https://github.com/bramstone/density-dependence-qSIP/tree/main/data>
}
\usage{
data(example_qsip)
}
\description{
A data set demonstrating the general organization and data necessary
 for the calculation of fractional enrichment via qSIP
}
\details{
\itemize{
\item asv_id. unique identifier for individual microbial taxa (here, amplicon sequence
 variants or ASVs) as identified by 16S amplicon sequencing
\item sampleID. unique identifier or code for experimental replicates
\item fraction. the fraction number of a replicate (1--22)
\item timepoint. the time of sampling for a replicate in days (0, 7)
\item isotope. factor of isotopes added to sample (values: NA, 16O, 18O)
\item iso_trt. factor specifying whether isotopes were added (light, label)
\item ecosystem. the site where soil samples originate from Northern Arizona (MC = mixed conifer, GL = grassland)
\item treatment. factor of nutrient amendments added (control, C for glcuose, CN for glucose and ammonium)
\item rep. numeric replicate from each plot (1--4 for unlabeled samples, 1--3 for labeled samples)
\item Density.g.ml. the density of individual fractions in grams per milliliter (1.6509988--1.736474)
\item avg_16S_g_soil. the concentration of bacterial 16S copies of individual fractions expressed per g dry soil (47--803,489)
\item seq_abund. the number of sequence reads per taxon (1--20,186)
\item Kingdom. the assigned kingdom of a taxon (Archaea, Bacteria, Unassigned, or Eukaryota)
\item Phylum. the assigned phylum of a taxon
\item Class. the assigned class of a taxon
\item Order. the assigned order of a taxon
\item Family. the assigned family of a taxon
\item Genus. the assigned genus of a taxon
}
}
\references{
\url{https://doi.org/10.1007/s00442-023-05322-z}
}
\keyword{datasets}
