% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.taxa.R
\name{explore.filter.taxa}
\alias{explore.filter.taxa}
\title{Explore Filter Taxa}
\usage{
explore.filter.taxa(
  DATA,
  trt.code.1 = NULL,
  trt.code.2 = NULL,
  trt.refs = NULL,
  vars = c("taxon", "copies", "tube", "trt.code"),
  min.reps
)
}
\arguments{
\item{DATA}{DATA}

\item{trt.code.1}{trt.code.1}

\item{trt.code.2}{trt.code.2}

\item{trt.refs}{trt.refs}

\item{vars}{vars}

\item{min.reps}{min.reps}
}
\description{
Define a function to use to filter the data so that only taxa with an appropriate level of occurrence and replication among the tubes and treatments of the experiment are retained for further analysis:
}
