% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{explore_filters}
\alias{explore_filters}
\title{Explore different filtering levels}
\usage{
explore_filters(data, filters = data@qsip@filter)
}
\arguments{
\item{data}{\code{Phylosip}-class object to pull feature taxa table from.}

\item{filters}{Optional data frame specifying filtering levels to explore.
The default is to utilize specifications from \code{data@qsip@filter}.
If \code{NULL}, \code{explore_filters} will generate statistics for all possible frequencies at both the replicate and fraction level.}
}
\value{
Some text here
}
\description{
Generates an S4 \code{filter} class object which stores filtering statistics
}
\details{
Some text here
}
\examples{
# Filter a tax table

}
\seealso{
\code{\link{create_filters}}
}
