% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_qsip}
\alias{filter_qsip}
\title{Filter taxa}
\usage{
filter_qsip(data, replicate = 0, fraction = 0, code = character())
}
\arguments{
\item{data}{\code{Phylosip}-class object to pull feature taxa table from.}

\item{replicate}{Numeric vector specifying the minimum frequency of occurrence of microbial taxa across replicates.
Keeping the default value of \code{0} will apply no frequency threshold at the replicate level}

\item{fraction}{Numeric vector specifying the minimum frequency of occurrence of microbial taxa across fractions within a sample.
Keeping the default value of \code{0} will apply no frequency threshold at the fraction level}

\item{code}{Optional character vector specifying a particular combination of replicate and fraction frequency to test.
Replicate and frequency combinations should be specified by separation with \code{:} (\emph{e.g.}, \code{'3:12'})}
}
\value{
Returns a
}
\description{
Filters taxa from qSIP portion of phyloseq object
}
\details{
\code{impose_filter} is primarily utilized within other functions.
}
\examples{
# Filter a phyloseq object

}
\seealso{
\code{\link{create_filters}}, \code{\link{explore_filters}}
}
