% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.taxa.R
\name{filter_taxa}
\alias{filter_taxa}
\title{Filter Taxa}
\usage{
filter_taxa(
  df,
  treatment1 = NULL,
  treatment2 = NULL,
  treatment_refs = NULL,
  min_reps = 3,
  taxon_column = "taxon",
  treatment_column = "unique.tmt",
  filter_column = "t.copies.ul",
  tube_column = "unique.tube"
)
}
\arguments{
\item{df}{Dataframe}

\item{treatment1}{trt.code.1}

\item{treatment2}{trt.code.2}

\item{treatment_refs}{trt.refs}

\item{min_reps}{min.reps}

\item{taxon_column}{Column header with taxon identifier (formerly vars[1])}

\item{treatment_column}{Column header with treatment identifier (formerly vars[4])}

\item{filter_column}{Column header with abundance or copy numbers (formerly vars[2])}

\item{tube_column}{Column header with tube or sample identifier (formerly vars[3])}
}
\value{
A list with filtered data and extras
\itemize{
  \item df: The filtered dataframe
  \item plot: A summary ggplot object
  \item treatment1: The argument passed to treatment1
  \item treatment2: The argument passed to treatment2
  \item treatment_refs: The argument passed to treatment_refs
  \item min_reps: The argument passed to min_reps
  \item taxon_column: The argument passed to taxon_column
  \item treatment_column: The argument passed to treatment_column
  \item filter_column: The argument passed to filter_column
  \item tube_column: The argument passed to tube_column
}
}
\description{
A function to filter the data so that only taxa with an appropriate level of
occurrence and replication among the tubes and treatments of the experiment
are retained for further analysis.
}
