% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{impose_filter}
\alias{impose_filter}
\title{Creates filtering criteria}
\usage{
impose_filter(data, replicate = 0, fraction = 0)
}
\arguments{
\item{data}{\code{Phylosip}-class object to pull feature taxa table from.}

\item{replicate}{Numeric vector specifying the minimum frequency of occurrence of microbial taxa across replicates.
Keeping the default value of \code{0} will apply no frequency threshold at the replicate level}

\item{fraction}{Numeric vector specifying the minimum frequency of occurrence of microbial taxa across fractions within a sample.
Keeping the default value of \code{0} will apply no frequency threshold at the fraction level}
}
\value{
Returns a
}
\description{
Returns binary vector specifying whether taxa from a phylosip object satisfy minimum frequency specifications
}
\details{
\code{impose_filter} is primarily utilized within other functions and imposes a hard filter on the data.
}
\examples{
# Filter a tax table
}
\seealso{
\code{\link{create_filters}}
}
