% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_init_methods.R
\docType{methods}
\name{initialize,qsip-method}
\alias{initialize,qsip-method}
\title{The S4 method for initializing a qsip object}
\usage{
\S4method{initialize}{qsip}(.Object, density = character(0),
  abund = character(0), rep_id = character(0),
  rep_group = character(0), iso = character(0),
  timepoint = character(0), filter = NULL)
}
\description{
The S4 method for initializing a qsip object
}
