% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtl.R
\name{mtl}
\alias{mtl}
\title{Matrix to long-form data}
\usage{
mtl(data)
}
\arguments{
\item{data}{An matrix of abundances, relativized or not, of taxonomic features across samples.
\code{data} may be an S3 or S4 class from the \code{Matrix} package but must be coercible to a \code{TsparseMatrix}.}
}
\value{
Returns a long-form data.table object where each row represents the abundance of a taxonomic feature
 in a sample.
}
\description{
Converts sample-by-feature matrix to long-form data.table
}
\details{
This function is included as a convenience for quickly transforming wide-format feature
 tables into long-format.

 Sometimes operations involving the \code{Matrix}-package classes begin to fail without explanation.
 When this happens, it is prudent to restart R in a fresh session.
}
\examples{
to go here

}
