% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MW_calc.R
\name{MW_calc}
\alias{MW_calc}
\title{Calculate Molecular Weight}
\usage{
MW_calc(
  df,
  density_column = "Density",
  copies_ul_column = "t.copies.ul",
  tube_column = "unique.tube"
)
}
\arguments{
\item{df}{data frame that includes x values (e.g., density of DNA) and y values (e.g., number of copies) and the replicate IDs (e.g., tube number)}

\item{density_column}{Column header with fraction density (formerly vars[1])}

\item{copies_ul_column}{Column header with abundance or copy numbers (formerly vars[2])}

\item{tube_column}{Column header with tube or sample identifier (formerly vars[3])}
}
\value{
A data.frame with the following columns
\itemize{
  \item WAD: mean of observed weighted average density for each replicate
  \item GC: guanine + cytosine content of the taxon's DNA
  \item MW: molecular weight (g/mol) of the taxon's DNA
  \item Catoms: mean number of carbon atoms per nucleotide for the taxon
  \item Oatoms: mean number of carbon atoms per nucleotide for the taxon
  \item Natoms: mean number of nitrogen atoms per nucleotide for the taxon
  \item message: warning message listing replicates that had no occurrences (i.e., all y values for a rep are 0); value for message is "none" when all replicates are present
}
}
\description{
Given a data frame for unlabeled treatment(s) for a specific taxon (i.e., no
added heavy isotopes), calculate the weighted average density (WAD), GC
content, molecular weight, and carbon content
}
\details{
assumes a fixed GC content, molecular weight, carbon, and nitrogen content
for each taxon based on the mean observed WAD
oxygen content is fixed for DNA regardless of GC content
the weighted average density of unlabeled DNA is a good proxy for GC content of the taxon, according to the well-known relationship between GC content and the density of DNA (uses equation from McHugh & Morrissey unpublished data)
(NOTE: This is not caused by the effect of GC content on molecular weight, which is negligible. It is caused by differences in the relationship between base composition and binding with water that occurs in CsCl.)

Written by Ben Koch & Natasja van Gestel
}
\note{
requires the function 'WAD_func'
}
