% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{parse_treatments}
\alias{parse_treatments}
\title{Parse Treatments}
\usage{
parse_treatments(treatment, sep = ";")
}
\arguments{
\item{treatment}{The treatment string}

\item{sep}{The delimiter}
}
\description{
A function to parse the treatment string in a qSIP dataset to a vector of
treatments with whitespace removed.
}
