% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/set_class.R
=======
% Please edit documentation in R/aa_set_class.R
>>>>>>> 7ea24d631be1eb62c29b2351469ed45b555f0930
\docType{class}
\name{phylosip-class}
\alias{phylosip-class}
\title{The S4 class for extending phyloseq objects with qSIP related data}
\description{
The S4 class for extending phyloseq objects with qSIP related data
}
