% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curve.R
\name{plot_curve}
\alias{plot_curve}
\title{Plot density curves}
\usage{
plot_curve(data, density = "character", abund = "character",
  iso_trt = "character", rep_id = "character",
  rep_group = "character", timepoint = "character",
  vertical = "logical", ...)
}
\arguments{
\item{data}{Data as a \code{phylosip} object}

\item{density}{Single length character matching to variable in \code{data} describing sample densities}

\item{abund}{Single length character matching to variable in \code{data} describing sample 16S gene abundance}

\item{iso_trt}{Single length character matching to variable in \code{data} used to distinguish between samples with and without isotope added.
The data this value references should be a factor with "light" treatments as the first level.}

\item{rep_id}{Single length character matching to variable in \code{data} describing sample/replicate names to summarize across fractions}

\item{rep_group}{Single length character matching to variable in \code{data} used to summarize values across groups of replicates}

\item{timepoint}{Single length character matching to variable in \code{data} used to distinguish between samples after certain incubation timepoints}
}
\value{
\code{plot_curve} returns a ggplot object that can be customized and added to using the common ggplot syntax.
}
\description{
Produces plot of density curves for data frame, phyloseq, or phylosip class objects using ggplot
}
\details{
Only the \code{data} argument is required if a \code{phylosip} object is provided. If \code{data} is of class \code{data.frame} or
  \code{phyloseq}, then the \code{density}, \code{abund}, and \code{iso_trt} arguments are required to plot density curves. Other arguments are
  required if replicates are to be grouped.
  Rows with values of \code{NA} in any column will be omitted.
}
\examples{
 # Load in example data

 # plot curve by specifying columns

 # specify qsip

 # plot curve

}
