% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/set_class.R
=======
% Please edit documentation in R/aa_set_class.R
>>>>>>> 7ea24d631be1eb62c29b2351469ed45b555f0930
\docType{class}
\name{qsip-class}
\alias{qsip-class}
\title{The S4 class for storing qSIP related data in list format}
\description{
Any resulting data will be stored in the \code{.Data} slot. Dscription of each argument found
 in \code{specify_qsip} function documentation.
}
