% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qsmelt.R
\name{qsmelt}
\alias{qsmelt}
\title{Melt qSIP data to long form}
\usage{
qsmelt(data, taxonomy = FALSE, abundance = FALSE, relativize = FALSE,
  exclude = c())
}
\arguments{
\item{data}{Data as a \code{phylosip} object}

\item{taxonomy}{Logical tag specifying whether or not to include taxonomic information of each feature. Default is \code{TRUE}.}

\item{abundance}{Logical tage specifying whether or not to include sequencing abundance of each feature. Default is \code{TRUE}.}

\item{relativize}{Logical tag specifying whether or not to present feature abundances as relative from 0 to 1 or to use those in @otu_table slot as is.}

\item{exclude}{Character vector specifying any qSIP measures to exclude from the returned data frame.}
}
\value{
\code{qsmelt} returns a single data frame with all qSIP related values for each taxon in each replicate, or group of replicates.
  Any replicate grouping (specified by \code{data@qsip@rep_group}) or distinct timepoints (specified by \code{data@qsip@timepoint}) will
  be represented in separate columns.
}
\description{
Converts data in phylosip object's qsip slot from several matrices to single, long-form data.frame
}
\details{
Details here on different specifications

  Note that WAD calculations will always be performed on a per-replicate basis. Thus, when mapped against grouped calculations, duplicate values
  will exist for non-WAD measures. To avoid this issue, exclude WAD calculations from the output with \code{exlude='wad'}.
}
\examples{
 # Load in example data

 # Calculate weighted average densities

 # Melt data

}
