% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_calc.R
\name{r_calc}
\alias{r_calc}
\title{Calculate R}
\usage{
r_calc(
  X.light,
  X.heavy,
  boot.out.light,
  boot.out.heavy,
  MW.out,
  iso.compare,
  days,
  var = "trt.code",
  growth.model = "exponential",
  prop.O.from.water = 0.5,
  CI = 0.9
)
}
\arguments{
\item{X.light}{data frame with data for treatment 1, the unlabeled, or 'light', treatment}

\item{X.heavy}{data frame with data for treatment 2, the labeled, or 'heavy', treatment}

\item{boot.out.light}{list containing output from boot.WAD.func or boot.TUBE.func for the 'light' treatment (corresponding to X.light); note that if a linear growth model is used, a list containing output from boot.TUBE.func is required}

\item{boot.out.heavy}{list containing output from boot.WAD.func or boot.TUBE.func for the 'heavy' treatment (corresponding to X.heavy); note that if a linear growth model is used, a list containing output from boot.TUBE.func is required}

\item{MW.out}{data frame containing output from MW.calc for the unlabeled treatment(s) for a taxon}

\item{iso.compare}{code specifying which isotopes are being compared; "13C" specifies the comparison between 13C (heavy) vs 12C (light) treatments; "18O" specifies the comparison between 18O (heavy) vs 16O (light) treatments; "15N" specifies the comparison between 15N (heavy) vs 14N (light) treatments; any other value returns an error}

\item{days}{duration of the incubation period in days}

\item{var}{variable name for the treatment ID columns in the data frames}

\item{growth.model}{model to use for calculating growth ("linear" or "exponential")}

\item{prop.O.from.water}{proportion of oxygen atoms in DNA that come from environmental water}

\item{CI}{confidence interval (0-1)}
}
\value{
A list with
\itemize{
\item boot.r: vector of bootstrapped estimates of intrinsic rate of increase (per day OR copies per day per g soil) based on the two treatments
\item obs.r: observed intrinsic rate of increase based on the two treatments
\item boot.r.mean: mean of the bootstrapped intrinsic rate of increase values
\item boot.r.median: median of the bootstrapped intrinsic rate of increase values
\item boot.r.CI: upper and lower confidence intervals (defined by argument 'CI') of the bootstrapped intrinsic rate of increase values
\item message: warning message listing replicates in either of the 2 treatments (does not include the reference group) that had no occurrences (i.e., all y values for a rep are 0); value for message is "none" when all replicates are present (single character value)
}
}
\description{
Given two treatments and the incubation period, returns an estimate (with uncertainty) of the intrinsic rate of increase (r, units for exponential model: per day; units for linear model: copies per day per g soil) based on the degree of labeling by the heavy isotope
}
\details{
Written by Ben Koch & Natasja van Gestel

Assumptions:
\itemize{
\item assumes either a linear growth model ( Nt=N0+(rt) ) or an exponential growth model ( Nt=N0*e^(rt) )
\item note that the units of r differ between the linear model (copies * day-1 * g soil-1) and the exponential model (1/day)
\item assumes a fixed GC content, molecular weight, carbon, and nitrogen content for each taxon based on the mean observed WAD
\item oxygen content is fixed for DNA regardless of GC content
\item the weighted average density of unlabeled DNA is a good proxy for GC content of the taxon, according to the well-known relationship between GC content and the density of DNA (Schildkraut et al. 1962)
\item (NOTE: This is not caused by the effect of GC content on molecular weight, which is negligible. It is caused by differences in the relationship between base composition and binding with water that occurs in CsCl.)
\item any change in weighted average density of the 16S DNA molecule is caused by incorporation of stable isotopes into the DNA
\item the relative change in weighted average density is equal to the relative change in molecular weight
\item (Incorporation of stable isotopes will not alter the binding properties between DNA and water in the CsCl gradient. Those properties are really important for affecting density, but they're a function of GC content, and should not change just because of heavy isotope incorporation.)
}
}
\note{
requires the functions 'WAD.func', 'boot.WAD.func', 'boot.TUBE.func', 'MW.calc', & 'comparison.message'

number of reps need not be equal between the two groups
}
