% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2.R
\name{read_qiime2_metadata}
\alias{read_qiime2_metadata}
\title{Read QIIME 2 data}
\usage{
read_qiime2_metadata(
  file,
  barcodes = FALSE,
  stringsAsFactors = TRUE,
  run_names = c()
)
}
\arguments{
\item{file}{The name of the file(s) which the data are to be read from. After ignoring any redundant headers, each row of the table appears
  as one line from the file(s). If there is no absolute path, the file name(s) will be relative to the current working directory, \code{getwd()}.
  Tilde-expansion is performed where supported.

  This can be a compressed file (see \code{\link[base]{connections}}).}

\item{barcodes}{Logical value indicating whether or not to keep barcodes and linker sequences.}

\item{stringsAsFactors}{Logical: should the character vector be converted to a factor? This tag excludes the first column of sample ID names and any possible
description columns (see details below).}

\item{run_names}{Character vector specifying an identifying name for each run to be read in. This will create a \code{Run} column in the resulting data frame}
}
\value{
\code{read_qiime2_metadata} returns a data frame from either a single or multiple QIIME2-formatted metadata tables.
}
\description{
Reads TSV or txt files used as QIIME 2 metadata
}
\details{
Some details
}
\examples{
to go here

}
