% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2.R
\name{read_qiime2_table}
\alias{read_qiime2_table}
\title{Read QIIME 2 data}
\usage{
read_qiime2_table(file, quiet = FALSE)
}
\arguments{
\item{file}{The name of the file which the feature data are to be read from. Each row of the table appears as one line of the file.
  If it does not contain an absolute path, the file name is relative to the current working directory, \code{getwd()}.
  Tilde-expansion is performed where supported.

  This can be a compressed file (see \code{\link[base]{file}}).}

\item{quiet}{Whether or not to produce messages notifying the user about file loading progress. Recommended for large tables
(larger than 300 Mb, uncompressed).}
}
\value{
\code{read_qiime2_table} returns an integer matrix of microbial features (ASVs or OTUs) along
  rows and samples along columns. Feature ID codes are stored as row names, and sample codes as column names.
}
\description{
Reads TSV tables of microbial features exported from QIIME 2's \code{biom} format
}
\details{
QIIME 2 uses the term features to refer to microbial taxonomic units in a way that is agnostic to the user's decision to use
  ASVs or OTUs. OTUs must first be clustered in QIIME 2, commonly at 97\% sequence similarity. ASVs represent unique sequences and so these tables will be
  larger than OTU tables for the same data set. QIIME 2 favors ASVs by default. For a discussion on the benefits of using ASVs over
  OTUs see: \url{https://www.nature.com/articles/ismej2017119} (referenced below).

  The input table is created using \code{qiime tools export} using a feature table \code{qza} to create a \code{biom} file
  followed by \code{biom convert} to produce a tab-delimited feature table. \code{read_qiime2_table} uses \code{\link{scan}} to
  read in the data in two parts, first the column headers, and then the rest of the data.
}
\examples{
data(example_data)

dat <- read_qiime2_table(data_name)

dim(dat)
dat[1:10,1:10]

#' @references
Callahan, B.J., P.J. McMurdie, S.P. Holmes. 2017. Exact sequence variants should replace operational
  taxonomic units in marker-gene data analysis. \\emph{ISME Journal} 11: 2639-2643.

}
\seealso{
\code{\link{read_qiime2_tax}}
}
