% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qiime2.R
\name{read_qiime2_tax}
\alias{read_qiime2_tax}
\title{Read QIIME 2 data}
\usage{
read_qiime2_tax(file, confidence = TRUE, feature_type = c("OTU",
  "ASV"), fill_na = TRUE)
}
\arguments{
\item{file}{The name of the file which the feature data are to be read from. Each row of the table appears as one line of the file.
  If it does not contain an absolute path, the file name is relative to the current working directory, \code{getwd()}.
  Tilde-expansion is performed where supported.

  This can be a compressed file (see \code{\link[base]{file}}).}

\item{feature_type}{Single character indicating whether the table is composed of ASVs (amplicon sequence variants) or
OTUs (operational taxonomic units). feature_type accepts only one argument.}

\item{fill_na}{Logical value indicating whether or not to return unclassified taxonomic values as blank (\code{fill_na=FALSE}) or NA.}
}
\value{
\code{read_qiime2_tax} returns a character matrix of taxonomic assignments from Kingdom to Species, identified to each
  microbial taxa by the row names of the matrix.
}
\description{
Reads TSV tables of microbial taxonomic information exported from QIIME 2's \code{biom} format
}
\details{
QIIME 2 uses the term features to refer to microbial taxonomic units in a way that is agnostic to the user's decision to use
  ASVs or OTUs. OTUs must first be clustered in QIIME 2, commonly at 97\% sequence similarity. ASVs represent unique sequences and so these tables will be
  larger than OTU tables for the same data set. QIIME 2 favors ASVs by default. For a discussion on the benefits of using ASVs over
  OTUs see: \url{https://www.nature.com/articles/ismej2017119} (referenced below).

  The input tab-separated TSV table is created using \code{qiime tools export} using a taxonomy \code{qza} file.
}
\examples{
data(example_data)

dat <- read_qiime2_tax(data_name, feature_type='otu', fill_na=TRUE)
head(dat)

dat <- read_qiime2_tax(data_name, feature_type='otu', fill_na=FALSE)
head(dat)

}
\references{
Callahan, B.J., P.J. McMurdie, S.P. Holmes. 2017. Exact sequence variants should replace operational
  taxonomic units in marker-gene data analysis. \emph{ISME Journal} 11: 2639-2643.
}
\seealso{
\code{\link{read_qiime2_table}}, \code{\link{specify_unclassifieds}}
}
