% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{sample_vec}
\alias{sample_vec}
\title{Create Sample Vector}
\usage{
sample_vec(x, ...)
}
\arguments{
\item{x}{vector of any length}
}
\value{
vector of length 'size' with elements drawn from x
}
\description{
Create a version of 'sample' that avoids the problem when length(x)==1
}
\details{
Written by Ben Koch & Natasja van Gestel
}
\note{
forces 'sample' to always treat x as a vector

(see examples on the ?sample page and also here: http://stackoverflow.com/questions/13990125/sampling-in-r-from-vector-of-varying-length)
}
