% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-methods.R
\docType{methods}
\name{show,qsip-method}
\alias{show,qsip-method}
\alias{show,phylosip-method}
\title{method extensions to show for phylosip objects.
Most code is taken directly from the phyloseq package}
\usage{
\S4method{show}{qsip}(object)

\S4method{show}{phylosip}(object)
}
\arguments{
\item{object}{Any R object}
}
\description{
See the general documentation of \code{\link[methods]{show}} method for
expected behavior.
}
\seealso{
\code{\link[methods]{show}}
}
