% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specify_qsip_method.R
\name{specify_qsip}
\alias{specify_qsip}
\title{Specify qSIP-related data to phylosip object}
\usage{
specify_qsip(data, density = "character", abund = "character",
  rep_id = "character", rep_num = "character",
  rep_group = "character", iso = "character", iso_trt = "character",
  timepoint = "character", filter_levels = "data.frame",
  filter = "character", ...)
}
\arguments{
\item{data}{Data as a \code{phyloseq} object}

\item{density}{Single length character matching to variable in \code{data} describing sample densities in UNITS}

\item{abund}{Single length character matching to variable in \code{data} describing sample 16S gene abundance}

\item{rep_id}{Single length character matching to variable in \code{data} describing sample/replicate names to summarize across fractions.
Required for calcualtion of weighted average DNA densities (\emph{per sample})}

\item{rep_num}{Single length character matching to variable in \code{data} matching replicates from the same sample origin across treatments
and any possible replicate groupings specified by \code{rep_group}.
Required if calculating \emph{per-capita} population rates of change where differences in abundance are only considered between sample from
the same origin}

\item{rep_group}{Single length character matching to variable in \code{data} used to summarize values across groups of replicates.
Required for calculation of atom excess fraction}

\item{iso}{Single length character vector describing which isotope was used, choices of \code{18O} or \code{13C} describing either
\eqn{^18^O}-labeled water or \eqn{^13^C}-labeled carbon inputs, respectively.
Required for calculation of molecular weights and atom excess fraction}

\item{iso_trt}{Single length character matching to variable in \code{data} used to distinguish between samples with and without isotope added.
The data this value references should be a facotr with "light" treatments as the first level.}

\item{timepoint}{Single length character matching to variable in \code{data} used to distinguish between samples after certain incubation
timepoints. The data that this value references should be a factor with time 0 values as the first level.
Required for calculation of population birth and death rates}

\item{filter}{Optional data frame of filtering options to apply to taxa based on sample and fraction-level frequency}

\item{...}{Arguments to change variables of existing \code{phylosip} object}
}
\value{
\code{specify_qsip} returns a \code{phyloseq}-like object of class \code{phylosip}.
  The \code{phylosip} object is an extension of \code{phyloseq} in that it contains an additional formal class,
  the \code{qsip} class. Objects of \code{qsip} class contain 7 slots and a \code{.Data} slot which is a list of
  all qSIP-related calculations (weighted average densities, molecular weights, and/or atom excess fraction) as well
  as any corresponding confidence intervals.
}
\description{
Creates or updates a phylosip object with experimental data and/or filtering criteria
}
\details{
Some details
}
\examples{
 # Create a new phylosip object from a phyloseq object

 # Update an existing phylosip object

}
