% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_test.R
\name{stress_test_plot}
\alias{stress_test_plot}
\title{Stress test plot}
\usage{
stress_test_plot(stress_test_results)
}
\arguments{
\item{stress_test_results}{Dataframe output of stress_test function}
}
\description{
Stress test plot
}
