% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavy.R
\name{tube_shift_heavy}
\alias{tube_shift_heavy}
\title{Calculates heavy isotope tube shift
Adds light and heavy tube shifts together}
\usage{
tube_shift_heavy(data, light, heavy_isotopes = c("13C"), low = 0, high = 0.1)
}
\arguments{
\item{data}{A dataframe containing taxon wad values per tube}

\item{light}{A dataframe containing light isotope adjusted taxon wad values averaged across all tubes}

\item{heavy_isotopes}{List of heavy isotopes}

\item{low}{Low}

\item{high}{High}
}
\description{
Calculates heavy isotope tube shift
Adds light and heavy tube shifts together
}
