% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/light.R
\name{tube_shift_light}
\alias{tube_shift_light}
\title{Calculates light isotope tube shift}
\usage{
tube_shift_light(data, light_isotopes = c("12C"))
}
\arguments{
\item{data}{A dataframe containing taxon wad values per tube}

\item{light_isotopes}{List of light isotopes}
}
\description{
Calculates light isotope tube shift
}
