% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wad_wide.R
\name{wad_wide}
\alias{wad_wide}
\title{Wide-format for WAD values}
\usage{
wad_wide(
  data,
  tax_id = c(),
  sample_id = c(),
  wads = "wad",
  iso_trt = c(),
  isotope = c(),
  average_light = TRUE
)
}
\arguments{
\item{data}{Data as a long-format data.table where each row represents a taxonomic feature within a single fraction.
Typically, this is the output from the \code{calc_wad} function.}

\item{tax_id}{Column name specifying unique identifier for each taxonomic feature.}

\item{sample_id}{Column name specifying unique identifier for each replicate.}

\item{wads}{Column name specifying weighted average density values.}

\item{iso_trt}{Column name specifying a two-level categorical column indicating whether a sample has been amended with a stable isotope (i.e., is "heavy") or if
isotopic composition is at natural abundance (i.e., "light").
Any terms may be applied but care should be taken for these values.
If supplied as a factor, \code{calc_excess} will take the lowest level as the "light" treatment and the higher
level as the "heavy" treatment.
Alternatively, if supplied as a character, \code{calc_excess} will coerce the column to a factor and with the default behavior wherein the
first value in alphabetical order will be assumed to be the lowest factor level (i.e. the "light" treatment).}

\item{average_light}{Whether to average the "light" WAD values.}
}
\value{
\code{wad_wide} returns a wide-transformed data.table where each row represents a taxonomic feature within a single replicate.
 If averaging unamended ("light") values is not performed, each taxon will have one numeric value and one \code{NA}
 value based on whether it was present in an unamended "light" sample or an amended "heavy" sample.
}
\description{
Transforms weighted average density values into wide-format
}
\examples{
 # Load in example data

 # Transform to wide format


}
\seealso{
\code{\link{calc_wad}}, \code{\link{calc_excess}}
}
