% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_d_wad.R
\name{calc_d_wad}
\alias{calc_d_wad}
\title{Calculation of weighted average density differences}
\usage{
calc_d_wad(data, filter = FALSE, return_diffs = FALSE,
  correction = FALSE, offset_taxa = 0.1, separate_light = FALSE,
  separate_label = TRUE, recalc = TRUE)
}
\arguments{
\item{data}{Data as a \code{phyloseq} object}

\item{filter}{Logical value specifying whether or not to filter taxa from the weighted average density calculation.
This will require \code{data} to have a filter applied with \code{\link{filter_qsip}}.}

\item{return_diffs}{Logical value specifying whether to return difference in mean WADs between light and heavy samples
or to return mean heavy WADs and mean light WADs (the default).}

\item{correction}{Logical value indicating whether or not to apply tube-level correction to labeled WAD values.}

\item{offset_taxa}{Value from 0 to 1 indicating the percentage of the taxa to utilize for calculating offset correction values.
Taxa are ordered by lowest difference in WAD values.
Default is \code{0.1} indicating 10 percent of taxa with the lowest difference in WAD values.}

\item{separate_light}{Logical value indicating whether or not WAD-light scores should be averaged across all replicate groups or not.
If \code{FALSE}, unlabeled WAD scores across all replicate groups will be averaged, creating a single molecular weight score per taxon
representing it's genetic molecular weight in the absence of isotope addition.}

\item{separate_label}{Logical value indicating whether or not WAD-label scores should be averaged across all replicate groups or not.
If \code{FALSE}, labeled WAD scores across all replicate groups will be averaged, creating a single molecular weight score per taxon
representing it's genetic molecular weight as a result of isotope addition. The default is \code{TRUE}.}

\item{recalc}{Logical value indicating whether or not to recalculate WAD values or use existing values. Default is \code{TRUE}.}
}
\value{
\code{calc_d_wad} adds two S4 Matrix objects to the \code{data@qsip@.Data} slot, one for differences
  in weighted average density, and the other for weighted average density values of light treatments only (to be used in
  future calculations). The row and column specifications will mirror those of the \code{phylosip}'s \code{\link{otu_table}},
  meaning if taxa are listed on the table rows, they will in the resulting S4 Matrix objects
}
\description{
Calculates differences in weighted average densities across replicates due to isotope incorporation
}
\details{
Some details about proper isotope control-treatment factoring. If weighted average densities have not been calculated
  beforehand, \code{calc_d_wad} will compute those first.

  More details about tube-level corrections.

  Unless a particular site is known to have significantly higher levels of naturally occurring isotopes, it is recommended to
  combine unlabeled (light) WAD scores in order to produce an average unlabeled WAD value that more accurately represents a
  taxon's natural WAD value.

  The equation for the difference in weighted average density of taxon \emph{i}, \eqn{\Delta W_{i}}, is given by:

  \deqn{\Delta W_{i} = W_{Lab,i} - W_{Light,i}}

  Where \eqn{W_{Lab,i}} indicates the weighted average density of taxon \emph{i} in heavy isotope-labeled treatment(s)
  and \eqn{W_{Light,i}} indicates that taxon's corresponding weighted average density in light, or unlabeled treatment(s)
}
\examples{
 # Load in example data


 # Calculate weighted average density differences

 @references
 Hungate, Bruce, \\emph{et al.} 2015. Quantitative microbial ecology through stable isotope probing.
 \\emph{Applied and Environmental Microbiology} \\strong{81}: 7570 - 7581.

}
\seealso{
\code{\link{calc_wad}}
}
