% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_excess.R
\name{calc_excess}
\alias{calc_excess}
\title{Calculation of atom excess}
\usage{
calc_excess(data, percent = FALSE, ci_method = c("", "bootstrap",
  "bayesian"), ci = 0.95, iters = 999, filter = FALSE,
  correction = FALSE, offset_taxa = 0.1, separate_light = FALSE,
  separate_label = TRUE)
}
\arguments{
\item{data}{Data as a \code{phyloseq} object}

\item{percent}{Logical value indicating whether or not to calculate atom percent excess (\code{percent=TRUE}) or atom excess fraction (the default)}

\item{ci_method}{Character value indicating how to calculate confidence intervals of stable isotope atom excess.
Options are \code{bootstrap} or \code{bayesian} (see \code{details} below for discussion on their differences).
The default is blank indicating that no confidence intervals will be calculated.}

\item{ci}{Numeric value from 0 to 1 indicating the width of the confidence interval for bootsrapped atom excess values.}

\item{iters}{Number of (subsampling) iterations to conduct to calculate confidence intervals. Default is \code{999}.}

\item{filter}{Logical vector specifying whether or not to filter taxa from the weighted average density calculation.
This will require \code{data} to have a filter applied with \code{\link{filter_qsip}}.}

\item{correction}{Logical value indicating whether or not to apply tube-level correction to labeled WAD values.}

\item{offset_taxa}{Value from 0 to 1 indicating the percentage of the taxa to utilize for calculating offset correction values.
Taxa are ordered by lowest difference in WAD values.
Default is \code{0.1} indicating 10 percent of taxa with the lowest difference in WAD values.}

\item{separate_light}{Logical value indicating whether or not WAD-light scores should be averaged across all replicate groups or not.
If \code{FALSE}, unlabeled WAD scores across all replicate groups will be averaged, creating a single molecular weight score per taxon
representing it's genetic molecular weight in the absence of isotope addition.}

\item{separate_label}{Logical value indicating whether or not WAD-label scores should be averaged across all replicate groups or not.
If \code{FALSE}, labeled WAD scores across all replicate groups will be averaged, creating a single molecular weight score per taxon
representing it's genetic molecular weight as a result of isotope addition. The default is \code{TRUE}.}
}
\value{
\code{calc_excess} adds an S4 Matrix class objects (which more efficiently stores sparse matrix data) to the \code{data@qsip@.Data} slot
  of molecular weights for each taxon at each group of replicates in the labeled and unlabeled groups. The row and column
  specifications will mirror those of the \code{phylosip}'s \code{\link{otu_table}}, meaning if taxa are listed on the table rows,
  they will in the resulting S4 Matrix class.

  Note that the bootstrap method produces a \emph{single} bootstrapped median (and matching confidence intervals) for groups of replicates,
  either grouped by isotope treatment alone, or also by some other grouping factor (if \code{data@qsip@rep_group} is specified).
  Using no bootstrap value allows separate enrichment values to be attained for each replicate, if \code{separate_label=TRUE}.
}
\description{
Calculates isotope incorporation in excess of natural abundances
}
\details{
Some details about proper isotope control-treatment factoring. If weighted average densities or the change in weighted average densities
  have not been calculated beforehand, \code{calc_mw} will compute those first.
}
\examples{
 # Load in example data

 # Calculate atom fraction excess

 # compare
 all.equal(aef*100, ape)

}
\seealso{
\code{\link{calc_wad}}, \code{\link{calc_d_wad}}, \code{\link{calc_mw}}
}
