% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_excess.R
\name{calc_excess}
\alias{calc_excess}
\title{Calculation of atom excess}
\usage{
calc_excess(data, percent = FALSE)
}
\arguments{
\item{data}{Data as a \code{phyloseq} object}

\item{percent}{Logical value indicating whether or not to calculate atom percent excess (\code{percent=TRUE}) or atom excess fraction (the default)}
}
\value{
\code{calc_excess} adds an S4 Matrix class objects (which more efficiently stores sparse matrix data) to the \code{data@qsip@.Data} slot
  of molecular weights for each taxon at each group of replicates in the labeled and unlabeled groups. The row and column
  specifications will mirror those of the \code{phylosip}'s \code{\link{otu_table}}, meaning if taxa are listed on the table rows,
  they will in the resulting S4 Matrix class.
}
\description{
Calculates isotope incorporation in excess of natural abundances
}
\details{
Some details about proper isotope control-treatment factoring. If weighted average densities or the change in weighted average densities
  have not been calculated beforehand, \code{calc_mw} will compute those first.
}
\examples{
 # Load in example data

 # Calculate atom fraction excess

 # compare
 all.equal(aef*100, ape)

}
\seealso{
\code{\link{calc_wad}}, \code{\link{calc_d_wad}}, \code{\link{calc_mw}}
}
