% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mw.R
\name{calc_mw}
\alias{calc_mw}
\title{Calculation of molecular weights}
\usage{
calc_mw(data, separate_wad_light = TRUE)
}
\arguments{
\item{data}{Data as a \code{phyloseq} object}

\item{separate_wad_light}{Logical value indicating whether or not WAD-light scores should be averaged across all replicate groups or not.
If \code{FALSE}, WAD scores across all replicate groups will be averaged, creating a single WAD score per taxon representing it's weighted
average density in the absence of isotope addition.}
}
\value{
\code{calc_mw} adds three S4 Matrix class objects (which more efficiently stores sparse matrix data) to the \code{data@qsip@.Data} slot
  of molecular weights for each taxon at each group of replicates in the labeled and unlabeled groups. The row and column
  specifications will mirror those of the \code{phylosip}'s \code{\link{otu_table}}, meaning if taxa are listed on the table rows,
  they will in the resulting S4 Matrix class.
}
\description{
Calculates molecular weights of microbial taxa due to isotope incorporation
}
\details{
Some details about proper isotope control-treatment factoring. If weighted average densities or the change in weighted average densities
  have not been calculated beforehand, \code{calc_mw} will compute those first.
}
\examples{
 # Load in example data

 # Calculate molecular weights

}
\seealso{
\code{\link{calc_wad}}, \code{\link{calc_d_wad}}
}
