% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pop.R
\name{calc_pop}
\alias{calc_pop}
\title{Calculation of population dynamics}
\usage{
calc_pop(data)
}
\arguments{
\item{data}{Data as a \code{phyloseq} object}
}
\value{
\code{calc_pop} adds two S4 Matrix class objects (which more efficiently stores sparse matrix data) to the \code{data@qsip@.Data} slot
  of population birth rates for each taxon at each group of replicates. The row and column specifications will mirror those of the \code{phylosip}'s
  \code{\link{otu_table}}, meaning if taxa are listed on the table rows, they will in the resulting S4 Matrix class.

  Timepoint should be in units of days, so that birth will be new 16S copies d-1 and death will be loss of light 16S copies d-1.
  Use of different time increments will yield growth rates (e.g. per hour), but must be appropriate for the frequency of sampling.
}
\description{
Calculates population growth and death rates
}
\details{
Some details about proper isotope control-treatment factoring and timepoint specification. If weighted average densities or the
  change in weighted average densities have not been calculated beforehand, \code{calc_pop} will compute those first.
}
\examples{
 # Load in example data

 # Calculate population fluxes

}
\seealso{
\code{\link{calc_mw}}
}
