% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_wad.R
\name{calc_wad}
\alias{calc_wad}
\title{Calculation of weighted average densities}
\usage{
calc_wad(data, filter = FALSE, pool_unlabeled = TRUE,
  calc_wvd = FALSE)
}
\arguments{
\item{data}{Data as a \code{phylosip} object}

\item{filter}{Logical vector specifying whether or not to filter taxa from the weighted average density calculation.
A hard filter is applied for \code{calc_wad} (meaning taxa are removed from all replicates/groups if they don't meet threshold).
Note, however, that this will have different outcomes depending on whether or not \code{filter_qsip} has been called
before calculations or not (see \code{details})}

\item{pool_unlabeled}{Logical vector specifying if unlabeled replicates should be pooled together across any grouping factor prior
to filtering. If \code{TRUE} (the default), unlabeled replicates will be pooled, and any soft filtering threshold will be applied to
\emph{all} unlabeled replicates together.}

\item{calc_wvd}{Logical vector specifying whether or not to calculate weighted variance of densities (WVD).
The weighted variance of densities of a taxon within a replicate is the complement to its weighted average.
It describes how much variance exists around every given weighted average density.
If \code{TRUE}, another output matrix (\code{wvd}) will be returned with identical dimensions to the weighted average
density output.}
}
\value{
\code{calc_wad} adds an S4 Matrix class (which more efficiently stores sparse matrix data) to the \code{.Data} slot within
  the \code{qsip} slot of the object. of weighted average density values for each taxon at each sample. The row and column
  specifications will mirror those of the \code{phylosip}'s \code{\link{otu_table}}, meaning if taxa are listed on the table rows,
  they will in the resulting S4 Matrix class.
}
\description{
Calculates weighted average densities for each microbial taxa in each sample replicate
}
\details{
Specifying \code{na.rm=TRUE} will allow \code{calc_wad} to calculate weighted average density values from samples
  that have one or more fractions without a valid density value. The default setting, \code{na.rm=FALSE}, returns values
  of \code{NA} for every taxa in a sample with missing density data.

  Filtering for \code{calc_wad} utilized \code{filter_qsip} and employs a hard filter. If filtering is specified using
  \code{filter=TRUE}, then filtering is employed on fractions \emph{only} (\emph{i.e.}, \code{filter_qsip} is implemented with
  \code{replicate=1}). In order to set more stringent hard filters, \code{filter_qsip} must be employed before calculating WAD values.

  The Weighted average density for taxon \emph{i} in replicate \emph{j}, \eqn{latex}{W_{ij}}, is:

  \deqn{latex}{W_{ij} = \sum_{k=1}^{K} x_{jk} \cdot \left( \frac{y_{ijk}}{y_{ij}} \right)}

  Where
  \eqn{latex}{y_{ij} = \sum_{k=1}^{K} y_{ijk}}: Total abundance of taxon \emph{i} in replicate \emph{j}, by summing across all fractions (\emph{K})

  \eqn{latex}{y_{ijk} = P_{ijk} \cdot f_{jk}}: The total abundance per \eqn{latex}{\mu}L for taxon \emph{i} in density fraction
  \emph{k} of replicate \emph{i} as calculated by it's relative abundance, \eqn{latex}{P_{ijk}} in that fraction multiplied by the
  total abundance of DNA or specific amplicons in that fraction, \eqn{latex}{f_{jk}}

  \eqn{latex}{x_{jk}}: Density of fraction \emph{k} of replicate \emph{j} in (g cm\eqn{latex}{^-3})
}
\examples{
 # Load in example data

 # Calculate weighted average densities

 @references
 Hungate, Bruce, \\emph{et al.} 2015. Quantitative microbial ecology through stable isotope probing.
 \\emph{Applied and Environmental Microbiology} \\strong{81}: 7570 - 7581.

}
\seealso{
\code{\link{filter_qsip}}
}
