% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_filters.R
\name{create_filters}
\alias{create_filters}
\title{Specify multiple taxa filting options for qSIP calculations}
\usage{
create_filters(replicate = 0, fraction = 0, rm_combn = character())
}
\arguments{
\item{replicate}{Numeric vector specifying the minimum frequency of occurrence of microbial taxa across replicates.
Keeping the default value of \code{0} will apply no frequency threshold at the replicate level}

\item{fraction}{Numeric vector specifying the minimum frequency of occurrence of microbial taxa across fractions within a sample.
Keeping the default value of \code{0} will apply no frequency threshold at the fraction level}

\item{rm_combn}{Optional character vector specifying any particular combinations of replicate and fraction frequency to remove.
Replicate and frequency combinations should be specified by separation with \code{:} (\emph{e.g.}, \code{'3:12'})}
}
\value{
\code{create_filters} produces a data frane of replicate frequencies and within-replicate-fraction frequencies to be
  investigated in downstream analyses.
}
\description{
Creates a data frame of per-replicate and per-fraction minimum incidence frequencies to retain microbial taxa in the dataset
}
\details{
\code{create_filters} should be used to specify the desired frequencies of microbial taxa \emph{prior} to calculation of
  atom excess fraction / atom percent excess. Filters must also be specified so that diagnostic functions can be run as well.
  Because \code{create_filters} uses \code{\link[base]{expand.grid}}, all possible combinations of the first two arguments will
  be used to construct the resulting data frame If certain combinations are not desirable, the \code{rm_combn} argument can be
  supplied to remove them.
}
\examples{
# Only filter on samples
 create_filters(2:3)

# Filter on samples and fractions
create_filters(2:3, 10:12)

# remove certain combinations
create_filters(2:3, 10:13, c('3:13', '2:10'))

}
