% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impose_filter.R
\name{impose_filter}
\alias{impose_filter}
\title{Filters taxa from data set}
\usage{
impose_filter(replicate = 0, fraction = 0, code = character(), data)
}
\arguments{
\item{replicate}{Numeric vector specifying the minimum frequency of occurrence of microbial taxa across replicates.
Keeping the default value of \code{0} will apply no frequency threshold at the replicate level}

\item{fraction}{Numeric vector specifying the minimum frequency of occurrence of microbial taxa across fractions within a sample.
Keeping the default value of \code{0} will apply no frequency threshold at the fraction level}

\item{code}{Optional character vector specifying a particular combination of replicate and fraction frequency to test.
Replicate and frequency combinations should be specified by separation with \code{:} (\emph{e.g.}, \code{'3:12'})}
}
\value{
Some text here
}
\description{
Filters taxa from a feature table matrix based on replicate and fraction frequencies
}
\details{
Some text here
}
\examples{
# Filter a tax table

}
\seealso{
\code{\link{create_filters}}
}
