% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_class.R
\docType{class}
\name{phylosip-class}
\alias{phylosip-class}
\title{The S4 class for extending phyloseq objects with qSIP related data}
\description{
The S4 class for extending phyloseq objects with qSIP related data
}
