% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_set_class.R
\docType{class}
\name{qsip-class}
\alias{qsip-class}
\title{The S4 class for storing qSIP related data in list format}
\description{
Any resulting data will be stored in the \code{.Data} slot. Dscription of each argument found
 in \code{specify_qsip} function documentation.
}
