% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_summary.R
\name{seq_summary}
\alias{seq_summary}
\title{Sequence and taxon count summaries}
\usage{
seq_summary(data, reads = c(), tax_id = c())
}
\arguments{
\item{data}{data.table object in long-format where each row represents a sequence feature in a given sequenced sample.}

\item{reads}{Column name specifying sequence read data.
Should be sequence read counts and not relativized abundances.}

\item{tax_id}{Column name specifying a unique identifier for each sequence feature.}
}
\value{
Produces formatted count of total sequence reads and unique taxonomic features across the data.
}
\description{
Returns the number of sequences and taxonomic features in the data. Useful for filtering.
}
\details{
QIIME 2 uses the term features to refer to microbial taxonomic units in a way that is agnostic to the user's decision to use
  ASVs or OTUs. OTUs must first be clustered in QIIME 2, commonly at 97\% sequence similarity. ASVs represent unique sequences and so these tables will be
  larger than OTU tables for the same data set. QIIME 2 favors ASVs by default. For a discussion on the benefits of using ASVs over
  OTUs see: \url{https://www.nature.com/articles/ismej2017119} (referenced below).
}
\examples{
data(example_data)

# initial sequence and ASV count? 
seq_summary(dat, 'seq_abund', 'taxon_id')

# Remove global singletons and doubletons
dat <- dat[seq_abund > 2]
seq_summary(dat, 'seq_abund', 'taxon_id')

# Identify lineages to remove
unassign <- tax[Kingdom == 'Unassigned', taxon_id]
euk <- tax[Kingdom == 'Eukaryota', taxon_id]
arch <- tax[Kingdom == 'Archaea', taxon_id]
mito_chloro <- tax[, tax_string := paste(Phylum, Class, Order, Family, Genus, Species, sep = ';')
                   ][grepl('mitochond|chloroplast', tax_string, ignore.case = T), taxon_id]

cat('Unassigned\n'); seq_summary(dat, 'seq_abund', 'taxon_id')
cat('Eukarya\n'); seq_summary(dat, 'seq_abund', 'taxon_id')
cat('Archaea\n'); seq_summary(dat, 'seq_abund', 'taxon_id')
cat('Mitochondria, Chloroplasts\n'); seq_summary(dat, 'seq_abund', 'taxon_id')

# Filter out lineages
dat <- dat[!taxon_id \%in\% c(arch, mito_chloro, euk, unassign)]
cat('\n\nFinal after filtering\n'); seq_summary(dat, 'seq_abund', 'taxon_id')

}
\seealso{
\code{\link{freq_filter}}
}
